workflow_system/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.0
google-generativeai==0.8.3
websockets==12.0
pydantic==2.10.3